feat: add piccolo ORM
Add piccolo ORM to connect to a database. I chose piccolo, as it
has a consise way of declaring tables with typehints and supports
both SQLite and Postgres, so it can be tested easily locally and on CI
without a full database but can have the full power of a production
database when needed.

I have currently just set it to use SQLite, and we can add Postgres
later when we have more information about how it will be deployed.