feat: Single Sign On login with Google
Add social sign in with Google using the authlib python library (https://docs.authlib.org/en/latest/index.html).

Instructions for how to set up with credentials are in the README.

There is currently no testing, as I don't know the best way to mock the external server needed as part of the handshake process.